+++ docs/Changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## v3.3.1
+* :core: Open up all logger functions
+* :core: Deprecate kauSwipeOnPostCreate and move functionality to onCreate
 * :searchview: Fix background tint
 
 ## v3.3.0